Split arithmetic routines from math.cc
