+++ chainerx_cc/chainerx/backward_test.cc
@@ -26,6 +26,7 @@
 #include "chainerx/graph.h"
 #include "chainerx/native/native_backend.h"
 #include "chainerx/op_node.h"
+#include "chainerx/routines/arithmetic.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
 #include "chainerx/shape.h"
