+++ chainerx_cc/chainerx/native/native_device/arithmetic.cc
@@ -7,10 +7,10 @@
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
 #include "chainerx/float16.h"
+#include "chainerx/kernels/arithmetic.h"
 #include "chainerx/kernels/math.h"
 #include "chainerx/native/elementwise.h"
 #include "chainerx/native/kernel_regist.h"
-#include "chainerx/routines/math.h"
 #include "chainerx/scalar.h"
 
 namespace chainerx {
