+++ chainerx_cc/chainerx/routines/indexing.cc
@@ -17,8 +17,8 @@
 #include "chainerx/constant.h"
 #include "chainerx/dtype.h"
 #include "chainerx/graph.h"
+#include "chainerx/kernels/arithmetic.h"
 #include "chainerx/kernels/indexing.h"
-#include "chainerx/kernels/math.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
